+++ tests/runnner.py
@@ -11,11 +11,11 @@ from zdevicemanager.base.base import cli
 
 # import your test modules
 
-from .test_data import DataTest
-from .test_jobs import JobsTest
+# from .test_data import DataTest
+# from .test_jobs import JobsTest
 from .test_conditions import ConditionsTest
-from .test_timestamp import TimeTest
-from .test_status import StatusTest
+# from .test_timestamp import TimeTest
+# from .test_status import StatusTest
 
 
 # if __name__ == '__main__':